+++ Lunar/DataModels/KbinPagination.swift
@@ -1,0 +1,12 @@
+//
+//  KbinPagination.swift
+//  Lunar
+//
+//  Created by Mani on 09/12/2023.
+//
+
+import Foundation
+
+struct KbinPagination: Codable {
+  let count, currentPage, maxPage, perPage: Int
+}