+++ Lunar/DataModels/KbinPostsModel.swift
@@ -1,0 +1,13 @@
+//
+//  KbinPostsModel.swift
+//  Lunar
+//
+//  Created by Mani on 09/12/2023.
+//
+
+import Foundation
+
+struct KbinPostsModel: Codable {
+  let posts: [KbinPostObject]
+  let pagination: KbinPagination
+}